+++ UI/src/components/utils/initializeTabHistory.js
@@ -1,0 +1,10 @@
+export const initializeTabHistory = (tab) => ({
+  ...tab,
+  history: [
+    {
+      nodes: tab.nodes || [],
+      edges: tab.edges || [],
+    },
+  ],
+  index: 0,
+});
